Update README.md
Question:
What is the difference between Tuples and ValueTuples?

Answer:
Tuples and ValueTuples are both data structures used to store multiple values, but they have some differences.

Tuples are immutable and can hold elements of different types. They are defined using parentheses and comma-separated values. Tuple elements can be accessed by their position (index). Tuples are part of the .NET Framework.

ValueTuples, on the other hand, are a newer addition to C# and are part of the System.ValueTuple namespace. They are also immutable, but they are specifically designed to hold a fixed number of elements (up to eight) and have performance optimizations. ValueTuples are defined using the ValueTuple struct and named properties for each element. ValueTuples support deconstruction, which allows extracting the individual elements of a tuple directly into separate variables.

In summary, Tuples are more flexible, allowing elements of different types and any number of elements, while ValueTuples are more performant and designed for a fixed number of elements.